third_party/blink/web_tests/fast/hidpi/image-srcset-invalid-descriptor-expected.txt

CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'w' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'w' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'h' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'h' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an 'h' descriptor and no 'w' descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an 'h' descriptor and no 'w' descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'w' descriptors or a mix of 'x' and 'w' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'w' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'w' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'h' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'h' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an 'h' descriptor and no 'w' descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has an 'h' descriptor and no 'w' descriptor.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("foo").clientWidth==200 is true
PASS document.getElementById("foo2").clientWidth==200 is true
PASS document.getElementById("foo3").clientWidth==200 is true
PASS document.getElementById("foo4").clientWidth==200 is true
PASS document.getElementById("foo5").clientWidth==200 is true
PASS document.getElementById("foo6").clientWidth==200 is true
PASS document.getElementById("foo7").clientWidth==200 is true
PASS document.getElementById("foo8").clientWidth==200 is true
PASS document.getElementById("foo9").clientWidth==200 is true
PASS document.getElementById("foo10").clientWidth==200 is true
This test passes if the image below is a 200px green square when the deviceScaleFactor is 1.
It tests that the 1x resource is ignored since it contains invalid descriptors.
The 2x resource has an intrinsic width of 200px, since it's a 400px wide image at 2x DPR.
         
